Address review comments
- Check ExtractionDelaySecs against ConsensusTimeout
- Pass GenerateCandidateMessage's context to Generate fn
- Add comment to ExtractionDelaySecs

+++ pkg/config/groups.go
@@ -138,7 +138,8 @@ type mempoolConfiguration struct {
 	// Enables mempool updates at startup
 	Updates updates
 
-	ExtractionDelaySecs uint32
+	// Artificial delay applied when mempool is empty
+	ExtractionDelaySecs int64
 }
 
 type updates struct {
